requirements.txt
asgiref==3.6.0
django==4.1.5
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
pillow==9.4.0